requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
httpx==0.25.1
python-telegram-bot==20.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
